+++ apparat/plugin_nautilus.py
@@ -29,27 +29,22 @@ def prepare_general(current_search_string, main_window):
     if current_search_string.startswith('!goto'):
         tools.debug_output('prepare_general', 'Case: Goto', 1)
         prepare_plugin_nautilus_goto(main_window)
-        return
 
     elif current_search_string == ('!recent'):
         tools.debug_output('prepare_general', 'Case: Recent', 1)
         prepare_plugin_nautilus_show_recent(main_window)
-        return
 
     elif current_search_string == ('!trash'):
         tools.debug_output('prepare_general', 'Case: Trash', 1)
         prepare_plugin_nautilus_open_trash(main_window)
-        return
 
     elif current_search_string == ('!network') or current_search_string == ('!net'):
         tools.debug_output('prepare_general', 'Case: Network', 1)
         prepare_plugin_nautilus_show_network_devices(main_window)
-        return
 
     else:
         tools.debug_output('prepare_general', 'Error: Unexpected nautilus plugin command', 3)
         main_window.status_notification_display_error('Unexpected nautilus plugin command')
-        return
 
     tools.debug_output('prepare_general', 'finished')