+++ apparat/prefs.py
@@ -7,7 +7,6 @@
 
 ## built-in modules
 import wx
-#import wx.grid
 
 ## apparat
 import constants
@@ -150,88 +149,131 @@ class UITabPluginCommands(wx.Panel):
         """Inits the plugin-commands tab"""
         wx.Panel.__init__(self, parent)
 
-        plugin_info = wx.StaticText(self, -1, "The following plugins exist so far (can't be disabled):", (20, 20))
+        ## Plugin: core (can not be disabled)
+        self.cb_enable_plugin_core = wx.CheckBox(self, -1, 'Core', (20, 60))
+        self.cb_enable_plugin_core.SetLabel('core')
+        self.cb_enable_plugin_core.SetToolTipString(u'Offers !help !prefs and !preferences.Plugin can not be disabled.')
+        self.cb_enable_plugin_core.SetValue(True)
+        self.cb_enable_plugin_core.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_click) # changing the checkbox change
 
         ## Plugin: Local search
-        cb_enable_plugin_local_search = wx.CheckBox(self, -1, 'Local-Search', (20, 60))
-        cb_enable_plugin_local_search.SetToolTipString(u'Enables search for files and folders in users home directory')
-        cb_enable_plugin_local_search.SetValue(True)
-        cb_enable_plugin_local_search.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_click) # Prevent changing state of enabled checkbox
+        self.cb_enable_plugin_local_search = wx.CheckBox(self, -1, 'Local-Search', (20, 60))
+        self.cb_enable_plugin_local_search.SetLabel('search_local')
+        self.cb_enable_plugin_local_search.SetToolTipString(u'Enables search for files and folders in users home directory')
+        cur_ini_value_for_plugin_search_local = ini.read_single_value('Plugins', 'enable_plugin_search_local')          # get current value from ini
+        if cur_ini_value_for_plugin_search_local == 'True':
+            self.cb_enable_plugin_local_search.SetValue(True)
+        else:
+            self.cb_enable_plugin_local_search.SetValue(False)
+        self.cb_enable_plugin_local_search.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_change) # changing the checkbox change
 
         ## Plugin: Internet search
-        cb_enable_plugin_insernet_search = wx.CheckBox(self, -1, 'Internet-Search', (20, 60))
-        cb_enable_plugin_insernet_search.SetToolTipString(u'Enables search for several popular web-services')
-        cb_enable_plugin_insernet_search.SetValue(True)
-        cb_enable_plugin_insernet_search.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_click) # Prevent changing state of enabled checkbox
+        self.cb_enable_plugin_internet_search = wx.CheckBox(self, -1, 'Internet-Search', (20, 60))
+        self.cb_enable_plugin_internet_search.SetLabel('search_internet')
+        self.cb_enable_plugin_internet_search.SetToolTipString(u'Enables search for several popular web-services')
+        cur_ini_value_for_plugin_search_internet = ini.read_single_value('Plugins', 'enable_plugin_search_internet')          # get current value from ini
+        if cur_ini_value_for_plugin_search_internet == 'True':
+            self.cb_enable_plugin_internet_search.SetValue(True)
+        else:
+            self.cb_enable_plugin_internet_search.SetValue(False)
+        self.cb_enable_plugin_internet_search.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_change) # changing the checkbox change
 
         ## Plugin: Nautilus
-        cb_enable_plugin_nautilus = wx.CheckBox(self, -1, 'Nautilus', (20, 60))
-        cb_enable_plugin_nautilus.SetToolTipString(u'Enables quick access to some nautilus locations/places')
-        cb_enable_plugin_nautilus.SetValue(True)
-        cb_enable_plugin_nautilus.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_click) # Prevent changing state of enabled checkbox
+        self.cb_enable_plugin_nautilus = wx.CheckBox(self, -1, 'Nautilus', (20, 60))
+        self.cb_enable_plugin_nautilus.SetLabel('nautilus')
+        self.cb_enable_plugin_nautilus.SetToolTipString(u'Enables quick access to some nautilus locations/places')
+        cur_ini_value_for_plugin_nautilus = ini.read_single_value('Plugins', 'enable_plugin_nautilus')          # get current value from ini
+        if cur_ini_value_for_plugin_nautilus == 'True':
+            self.cb_enable_plugin_nautilus.SetValue(True)
+        else:
+            self.cb_enable_plugin_nautilus.SetValue(False)
+        self.cb_enable_plugin_nautilus.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_change) # changing the checkbox change
 
         ## Plugin: Screenshot
-        cb_enable_plugin_screenshot = wx.CheckBox(self, -1, 'Screenshot', (20, 60))
-        cb_enable_plugin_screenshot.SetToolTipString(u'Enables simple screenshot functions')
-        cb_enable_plugin_screenshot.SetValue(True)
-        cb_enable_plugin_screenshot.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_click) # Prevent changing state of enabled checkbox
+        self.cb_enable_plugin_screenshot = wx.CheckBox(self, -1, 'Screenshot', (20, 60))
+        self.cb_enable_plugin_screenshot.SetLabel('screenshot')
+        self.cb_enable_plugin_screenshot.SetToolTipString(u'Enables simple screenshot functions')
+        cur_ini_value_for_plugin_screenshot = ini.read_single_value('Plugins', 'enable_plugin_screenshot')          # get current value from ini
+        if cur_ini_value_for_plugin_screenshot == 'True':
+            self.cb_enable_plugin_screenshot.SetValue(True)
+        else:
+            self.cb_enable_plugin_screenshot.SetValue(False)
+        self.cb_enable_plugin_screenshot.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_change) # changing the checkbox change
 
         ## Plugin: Session
-        cb_enable_plugin_session = wx.CheckBox(self, -1, 'Session', (20, 60))
-        cb_enable_plugin_session.SetToolTipString(u'Enables several session commands')
-        cb_enable_plugin_session.SetValue(True)
-        cb_enable_plugin_session.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_click) # Prevent changing state of enabled checkbox
+        self.cb_enable_plugin_session = wx.CheckBox(self, -1, 'Session', (20, 60))
+        self.cb_enable_plugin_session.SetLabel('session')
+        self.cb_enable_plugin_session.SetToolTipString(u'Enables several session commands')
+        cur_ini_value_for_plugin_session = ini.read_single_value('Plugins', 'enable_plugin_session')          # get current value from ini
+        if cur_ini_value_for_plugin_session == 'True':
+            self.cb_enable_plugin_session.SetValue(True)
+        else:
+            self.cb_enable_plugin_session.SetValue(False)
+        self.cb_enable_plugin_session.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_change) # changing the checkbox change
 
         ## Plugin: Shell
-        cb_enable_plugin_shell = wx.CheckBox(self, -1, 'Shell', (20, 60))
-        cb_enable_plugin_shell.SetToolTipString(u'Enable executing shell commands')
-        cb_enable_plugin_shell.SetValue(True)
-        cb_enable_plugin_shell.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_click) # Prevent changing state of enabled checkbox
+        self.cb_enable_plugin_shell = wx.CheckBox(self, -1, 'Shell', (20, 60))
+        self.cb_enable_plugin_shell.SetLabel('shell')
+        self.cb_enable_plugin_shell.SetToolTipString(u'Enable executing shell commands')
+        cur_ini_value_for_plugin_shell = ini.read_single_value('Plugins', 'enable_plugin_shell')          # get current value from ini
+        if cur_ini_value_for_plugin_shell == 'True':
+            self.cb_enable_plugin_shell.SetValue(True)
+        else:
+            self.cb_enable_plugin_shell.SetValue(False)
+        self.cb_enable_plugin_shell.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_change) # changing the checkbox change
 
         ## Plugin: Misc
-        cb_enable_plugin_misc = wx.CheckBox(self, -1, 'Misc', (20, 60))
-        cb_enable_plugin_misc.SetToolTipString(u'Enable other stuff')
-        cb_enable_plugin_misc.SetValue(True)
-        cb_enable_plugin_misc.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_click) # Prevent changing state of enabled checkbox
+        self.cb_enable_plugin_misc = wx.CheckBox(self, -1, 'Misc', (20, 60))
+        self.cb_enable_plugin_misc.SetToolTipString(u'Enable other stuff')
+        self.cb_enable_plugin_misc.SetLabel('misc')
+        cur_ini_value_for_plugin_misc = ini.read_single_value('Plugins', 'enable_plugin_misc')          # get current value from ini
+        if cur_ini_value_for_plugin_misc == 'True':
+            self.cb_enable_plugin_misc.SetValue(True)
+        else:
+            self.cb_enable_plugin_misc.SetValue(False)
+        self.cb_enable_plugin_misc.Bind(wx.EVT_CHECKBOX, self.on_plugin_checkbox_change) # changing the checkbox change
 
         ## Link to plugin commands description
         wxHyperlinkCtrl = wx.HyperlinkCtrl(self, -1, 'Plugin command details', constants.APP_URL+'#plugins')
 
+
+
+        ## -----------------------
         ## Layout
-        ##
+        ## -----------------------
         pref_sizer = wx.BoxSizer(wx.VERTICAL) # define layout container
         pref_sizer.AddSpacer(10)
 
-        ## general info
-        pref_sizer.Add(plugin_info, 0, wx.ALL, border=10)
+        ## core
+        pref_sizer.Add(self.cb_enable_plugin_core, 0, wx.ALL, border=10)
         pref_sizer.AddSpacer(5)
 
-        ## Local search
-        pref_sizer.Add(cb_enable_plugin_local_search, 0, wx.ALL, border=10)
-        pref_sizer.AddSpacer(5)
-
-        ## Internet search
-        pref_sizer.Add(cb_enable_plugin_insernet_search, 0, wx.ALL, border=10)
+        ## misc
+        pref_sizer.Add(self.cb_enable_plugin_misc, 0, wx.ALL, border=10)
         pref_sizer.AddSpacer(5)
 
         ## Nautilus
-        pref_sizer.Add(cb_enable_plugin_nautilus, 0, wx.ALL, border=10)
+        pref_sizer.Add(self.cb_enable_plugin_nautilus, 0, wx.ALL, border=10)
+        pref_sizer.AddSpacer(5)
+
+        ## screenshot
+        pref_sizer.Add(self.cb_enable_plugin_screenshot, 0, wx.ALL, border=10)
+        pref_sizer.AddSpacer(5)
+
+        ## search internet
+        pref_sizer.Add(self.cb_enable_plugin_internet_search, 0, wx.ALL, border=10)
+        pref_sizer.AddSpacer(5)
+
+        ## search local
+        pref_sizer.Add(self.cb_enable_plugin_local_search, 0, wx.ALL, border=10)
         pref_sizer.AddSpacer(5)
 
         ## Session
-        pref_sizer.Add(cb_enable_plugin_screenshot, 0, wx.ALL, border=10)
-        pref_sizer.AddSpacer(5)
-
-        ## Session
-        pref_sizer.Add(cb_enable_plugin_session, 0, wx.ALL, border=10)
+        pref_sizer.Add(self.cb_enable_plugin_session, 0, wx.ALL, border=10)
         pref_sizer.AddSpacer(5)
 
         ## Shell
-        pref_sizer.Add(cb_enable_plugin_shell, 0, wx.ALL, border=10)
-        pref_sizer.AddSpacer(5)
-
-        ## Misc
-        pref_sizer.Add(cb_enable_plugin_misc, 0, wx.ALL, border=10)
+        pref_sizer.Add(self.cb_enable_plugin_shell, 0, wx.ALL, border=10)
         pref_sizer.AddSpacer(20)
 
         ## Hyperlink to docs
@@ -244,4 +286,17 @@ class UITabPluginCommands(wx.Panel):
         """Handle plugin checkbox click - prevent changing values"""
         e_obj = evt.GetEventObject()
         e_obj.SetValue(not e_obj.GetValue()) # plugin checkboxes are always TRUE/checked
-        wx.MessageBox(constants.APP_NAME+' plugins can not be disabled', 'Error', wx.OK | wx.ICON_WARNING)
+        wx.MessageBox('Plugin can not be disabled', 'Error', wx.OK | wx.ICON_WARNING)
+
+
+    def on_plugin_checkbox_change(self, event):
+        """Triggered if a checkbox state is changing and writes the new state to ini"""
+        btn = event.GetEventObject().GetLabel()
+        value = event.GetEventObject().GetValue()
+
+        if value is True:
+            tools.debug_output('on_plugin_checkbox_change', 'Enabled Plugin: '+btn, 1)
+            ini.write_single_value('Plugins', 'enable_plugin_'+btn, 'True')
+        else:
+            tools.debug_output('on_plugin_checkbox_change', 'Disabled Plugin: '+btn, 1)
+            ini.write_single_value('Plugins', 'enable_plugin_'+btn, 'False')